Fix bug in sequence value utility where it is actually using Integer instead of Long, so very large sequence values cannot be generated
git-svn-id: https://svn.apache.org/repos/asf/incubator/ofbiz/trunk@469283 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/entity/src/org/ofbiz/entity/util/SequenceUtil.java
@@ -191,7 +191,7 @@ public class SequenceUtil {
                             rs = stmt.executeQuery(sql);
                             boolean gotVal1 = false;
                             if (rs.next()) {
-                                val1 = rs.getInt(parentUtil.idColName);
+                                val1 = rs.getLong(parentUtil.idColName);
                                 gotVal1 = true;
                             }
                             rs.close();
@@ -214,7 +214,7 @@ public class SequenceUtil {
                             rs = stmt.executeQuery(sql);
                             boolean gotVal2 = false;
                             if (rs.next()) {
-                                val2 = rs.getInt(parentUtil.idColName);
+                                val2 = rs.getLong(parentUtil.idColName);
                                 gotVal2 = true;
                             }
